First working implementation.
