+++ src/sbt-test/debian/simple-jdeb/build.sbt
@@ -1,0 +1,22 @@
+import NativePackagerKeys._
+
+packagerSettings
+
+mapGenericFilesToLinux
+
+name := "debian-test"
+
+version := "0.1.0"
+
+maintainer := "Josh Suereth <joshua.suereth@typesafe.com>"
+
+packageSummary := "Test debian package"
+
+packageDescription := """A fun package description of our software,
+  with multiple lines."""
+
+debianPackageDependencies in Debian ++= Seq("java2-runtime", "bash (>= 2.05a-11)")
+
+debianPackageRecommends in Debian += "git"
+
+packageBin in Debian <<= debianJDebPackaging in Debian
