+++ src/sbt-test/debian/simple-jdeb/test
@@ -1,0 +1,3 @@
+# Run the debian packaging.
+> debian:package-bin
+$ exists target/debian-test_0.1.0_all.deb
